requirements.txt
fastapi
uvicorn
pandas
numpy
cohere
openai
scikit-learn
nltk
pymongo
python-dotenv
kafka-python
annoy